requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.7
django-axes==6.1.1
django-cors-headers==4.3.1
django-js-asset==2.1.0
django-mptt==0.15.0
django-oauth-toolkit==2.3.0
django-recaptcha==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
drf-spectacular==0.26.5
ecdsa==0.18.0
factory-boy==3.3.0
Faker==19.13.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.6.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
